Create text_alignment.py
I truly hate this problem. VERY bad description yet TRULY hard to count all of the empty slots and 'H' slots. I hope my code can be helpful to those who struggled to understand what is going on with the given example. I tried to create functions for tasks such as creating the triangles or the H logo itself in order to make the code more readable since it's already very hard to understand the problem itself.